ISSUE #1108 Disable NFC scan from phone home screen, make it possible to scan tag only from dashboard or force claim screens
