+++ app/src/main/java/com/ruuvi/station/network/ui/claim/ClaimSensorActivity.kt
@@ -1,12 +1,7 @@
 package com.ruuvi.station.network.ui.claim
 
-import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
-import android.nfc.NfcAdapter
-import android.nfc.tech.NfcA
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.compose.animation.*
@@ -31,7 +26,7 @@ import com.ruuvi.station.app.ui.components.*
 import com.ruuvi.station.app.ui.theme.RuuviStationTheme
 import com.ruuvi.station.app.ui.theme.RuuviTheme
 import com.ruuvi.station.network.ui.SignInActivity
-import com.ruuvi.station.nfc.NfcScanReciever
+import com.ruuvi.station.util.base.NfcActivity
 import com.ruuvi.station.util.extensions.navigate
 import com.ruuvi.station.util.extensions.viewModel
 import org.kodein.di.Kodein
@@ -40,7 +35,7 @@ import org.kodein.di.android.closestKodein
 import timber.log.Timber
 
 @OptIn(ExperimentalAnimationApi::class)
-class ClaimSensorActivity : AppCompatActivity(), KodeinAware {
+class ClaimSensorActivity : NfcActivity(), KodeinAware {
 
     override val kodein: Kodein by closestKodein()
 
@@ -50,35 +45,9 @@ class ClaimSensorActivity : AppCompatActivity(), KodeinAware {
         }
     }
 
-    private var nfcAdapter: NfcAdapter? = null
-
-    val intentFiltersArray = arrayOf(
-        IntentFilter(NfcAdapter.ACTION_NDEF_DISCOVERED).apply {
-            try {
-                addDataType("text/plain")
-            } catch (e: IntentFilter.MalformedMimeTypeException) {
-                throw RuntimeException("fail", e)
-            }}
-    )
-
-    val techListsArray = arrayOf(arrayOf<String>(NfcA::class.java.name))
-
-    var nfcIntent : Intent? = null
-
-    var pendingIntent: PendingIntent? = null
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         WindowCompat.setDecorFitsSystemWindows(window, true)
-
-        nfcIntent = Intent(this, javaClass).apply {
-            addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
-        }
-
-        pendingIntent = PendingIntent.getActivity(this, 0, nfcIntent,
-            PendingIntent.FLAG_MUTABLE)
-
-        nfcAdapter = NfcAdapter.getDefaultAdapter(this)
 
         setContent {
             RuuviTheme {
@@ -87,24 +56,6 @@ class ClaimSensorActivity : AppCompatActivity(), KodeinAware {
         }
 
         viewModel.checkClaimState()
-    }
-
-    override fun onPause() {
-        super.onPause()
-        nfcAdapter?.disableForegroundDispatch(this)
-    }
-
-    override fun onResume() {
-        super.onResume()
-        nfcAdapter?.enableForegroundDispatch(this, pendingIntent, intentFiltersArray, techListsArray)
-
-    }
-
-    override fun onNewIntent(intent: Intent) {
-        super.onNewIntent(intent)
-        if (intent.type == "text/plain" && intent.action == NfcAdapter.ACTION_NDEF_DISCOVERED) {
-            NfcScanReciever.nfcScanned(intent)
-        }
     }
 
     @Composable
