+++ app/src/main/java/com/ruuvi/station/tagdetails/ui/SensorCardActivity.kt
@@ -4,13 +4,9 @@ import android.app.Activity
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
 import android.net.Uri
-import android.nfc.NfcAdapter
-import android.nfc.tech.NfcA
 import android.os.Bundle
 import androidx.activity.compose.setContent
-import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.animation.*
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
@@ -64,6 +60,7 @@ import com.ruuvi.station.tag.domain.isLowBattery
 import com.ruuvi.station.tagsettings.ui.TagSettingsActivity
 import com.ruuvi.station.units.domain.UnitsConverter
 import com.ruuvi.station.util.Days
+import com.ruuvi.station.util.base.NfcActivity
 import org.kodein.di.KodeinAware
 import org.kodein.di.android.closestKodein
 import com.ruuvi.station.util.extensions.*
@@ -73,7 +70,7 @@ import kotlinx.coroutines.flow.collectLatest
 import org.kodein.di.generic.instance
 import timber.log.Timber
 
-class SensorCardActivity : AppCompatActivity(), KodeinAware {
+class SensorCardActivity : NfcActivity(), KodeinAware {
 
     override val kodein by closestKodein()
 
@@ -86,35 +83,9 @@ class SensorCardActivity : AppCompatActivity(), KodeinAware {
         )
     }
 
-    private var nfcAdapter: NfcAdapter? = null
-
-    val intentFiltersArray = arrayOf(
-        IntentFilter(NfcAdapter.ACTION_NDEF_DISCOVERED).apply {
-            try {
-                addDataType("text/plain")
-            } catch (e: IntentFilter.MalformedMimeTypeException) {
-                throw RuntimeException("fail", e)
-            }}
-    )
-
-    val techListsArray = arrayOf(arrayOf<String>(NfcA::class.java.name))
-
-    var nfcIntent : Intent? = null
-
-    var pendingIntent: PendingIntent? = null
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        nfcIntent = Intent(this, javaClass).apply {
-            addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
-        }
-
-        pendingIntent = PendingIntent.getActivity(this, 0, nfcIntent,
-            PendingIntent.FLAG_MUTABLE)
-
-        nfcAdapter = NfcAdapter.getDefaultAdapter(this)
 
         setContent {
             RuuviTheme {
@@ -148,24 +119,6 @@ class SensorCardActivity : AppCompatActivity(), KodeinAware {
                     addSensor = viewModel::addSensor
                 )
             }
-        }
-    }
-
-    override fun onPause() {
-        super.onPause()
-        nfcAdapter?.disableForegroundDispatch(this)
-    }
-
-    override fun onResume() {
-        super.onResume()
-        nfcAdapter?.enableForegroundDispatch(this, pendingIntent, intentFiltersArray, techListsArray)
-
-    }
-
-    override fun onNewIntent(intent: Intent) {
-        super.onNewIntent(intent)
-        if (intent.type == "text/plain" && intent.action == NfcAdapter.ACTION_NDEF_DISCOVERED) {
-            NfcScanReciever.nfcScanned(intent)
         }
     }
